feat: 로드맵 풀백 게이트 + ExitPlan 수량 float 전환 + LLM 프롬프트 개선
- 로드맵 풀백 기술분석(analysis/technical) 및 전략(strategy) 모듈 신규 추가
- ExitPlan total_quantity float 마이그레이션으로 코인 소수 수량 지원
- Codex CLI 프로바이더 및 LLM 팩토리 개선, 분석/리뷰 프롬프트 업데이트
- TradingAgent mixin 전반 업데이트 (포트폴리오, 분석, 사이클, 이벤트)
- DecisionMaker·MarketScanner·EventDetector 로드맵 풀백 게이트 연동
- 관련 유닛테스트 추가 및 기존 테스트 갱신

+++ models/exit_plan.py
@@ -29,7 +29,7 @@ class ExitPlan(Base, TimestampMixin):
     market: Mapped[str] = mapped_column(String(10), nullable=False)  # KRX, NYSE, NASDAQ
 
     avg_entry_price: Mapped[float] = mapped_column(Float, nullable=False)
-    total_quantity: Mapped[int] = mapped_column(Integer, nullable=False)
+    total_quantity: Mapped[float] = mapped_column(Float, nullable=False)
 
     # 다단계 청산 레벨 (JSON string)
     levels: Mapped[str] = mapped_column(Text, nullable=False)
@@ -60,7 +60,7 @@ class ExitPlanHistory(Base):
     # INITIAL | ADD_ON_PYRAMID | ADD_ON_AVERAGE_DOWN | AI_REVIEW | TRAILING_UPDATE
 
     avg_entry_price: Mapped[float] = mapped_column(Float, nullable=False)
-    total_quantity: Mapped[int] = mapped_column(Integer, nullable=False)
+    total_quantity: Mapped[float] = mapped_column(Float, nullable=False)
     ai_reasoning: Mapped[str | None] = mapped_column(Text, nullable=True)
 
     created_at: Mapped[datetime] = mapped_column(DateTime, nullable=False)
